Compile and link rocketspeed servers on MacOS. Import libevent 2.0.21 that works for both Android and MacOS. Run rocketbench from MacOS.
Summary: rocketbench now runs succesfully from MacOS

Test Plan: ./rocketbench

Reviewers: stupaq-fb, pja

Reviewed By: pja

Differential Revision: https://reviews.facebook.net/D34107

+++ src/client/storage/file_storage.cc
@@ -332,7 +332,7 @@ void FileStorage::HandleSnapshotCommand(std::unique_ptr<Command> command) {
   for (auto& it : worker_data.topics_subscribed) {
     PutFixed64(&buffer, it.second.GetSequenceNumber());
     PutFixed16(&buffer, it.first.namespace_id);
-    PutFixed32(&buffer, it.first.topic_name.size());
+    PutFixed32(&buffer, static_cast<uint32_t>(it.first.topic_name.size()));
     buffer.append(it.first.topic_name);
   }
   // Append to snapshot state.
